# hosted-agent-kit/tests/contract/openapi.snapshot.json
{
  "components": {
    "schemas": {
...
      "AgentCapabilities": {
        "description": "What a caller can do with one agent. Contains no administrative detail.",
        "properties": {
          "agent_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Pinned version for new sessions, or null.",
            "title": "Agent Version"
          },
          "endpoints": {
            "items": {
              "type": "string"
            },
            "title": "Endpoints",
            "type": "array"
          },
          "idempotency_ttl_seconds": {
            "description": "How long a repeated Idempotency-Key replays the stored response. 0 is off.",
            "title": "Idempotency Ttl Seconds",
            "type": "number"
          },
          "limits": {
            "$ref": "#/components/schemas/Limits"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "protocol": {
            "description": "responses or invocations.",
            "title": "Protocol",
            "type": "string"
          },
          "queue_enabled": {
            "title": "Queue Enabled",
            "type": "boolean"
          },
          "queue_max_wait_seconds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Queue Max Wait Seconds"
          },
          "request_content_types": {
            "items": {
              "type": "string"
            },
            "title": "Request Content Types",
            "type": "array"
          },
          "stateful": {
            "description": "True when each user (and conversation) keeps a session.",
            "title": "Stateful",
            "type": "boolean"
          },
          "streaming": {
            "description": "`request`: the caller asks with `stream`. `agent`: the agent decides.",
            "title": "Streaming",
            "type": "string"
          },
          "supports_conversation_key": {
            "title": "Supports Conversation Key",
            "type": "boolean"
          }
        },
        "required": [
          "name",
          "protocol",
          "stateful",
          "streaming",
          "supports_conversation_key",
          "endpoints",
          "request_content_types",
          "queue_enabled",
          "queue_max_wait_seconds",
          "idempotency_ttl_seconds",
          "agent_version",
          "limits"
        ],
        "title": "AgentCapabilities",
        "type": "object"
      },
...
      "ProblemDetails": {
        "properties": {
          "correlation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Correlation Id"
          },
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "error_code": {
            "title": "Error Code",
            "type": "string"
          },
          "phase": {
            "description": "Where the request failed: request, auth, queue, create, invoke or stream.",
            "title": "Phase",
            "type": "string"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          },
          "retry_after_seconds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retry After Seconds"
          },
          "retry_safe": {
            "description": "True when retrying the same request cannot repeat work the agent already did.",
            "title": "Retry Safe",
            "type": "boolean"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "error_code",
          "phase",
          "retry_safe"
        ],
        "title": "ProblemDetails",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "bearerFormat": "JWT",
        "description": "Microsoft Entra ID access token. The user identity is the `oid` claim.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "BFF and worker-pool controller for Microsoft Foundry hosted agents.",
    "title": "Foundry Hosted Agent Pool",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/agents": {
      "get": {
        "operationId": "list_agents_v1_agents_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AgentCapabilities"
                  },
                  "title": "Response List Agents V1 Agents Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Content Too Large"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "502": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Bad Gateway"
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Service Unavailable"
          },
          "504": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Gateway Timeout"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List the agents a caller can use",
        "tags": [
          "agents"
        ]
      }
    },
    "/v1/agents/{agent_name}": {
      "get": {
        "operationId": "describe_agent_v1_agents__agent_name__get",
        "parameters": [
          {
            "description": "Configured agent",
            "in": "path",
            "name": "agent_name",
            "required": true,
            "schema": {
              "description": "Configured agent",
              "maxLength": 128,
              "pattern": "^[A-Za-z0-9][A-Za-z0-9_-]{0,127}$",
              "title": "Agent Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentCapabilities"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Content Too Large"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "502": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Bad Gateway"
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Service Unavailable"
          },
          "504": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Gateway Timeout"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Describe one agent: protocol, streaming, limits and endpoints",
        "tags": [
          "agents"
        ]
      }
    },
...
  }
}